refactor: EventCard 통일 (#131)
