+++ src/components/EventCard.jsx
@@ -3,6 +3,7 @@ import { useState } from 'react';
 import Swal from 'sweetalert2';
 import { Heart } from 'lucide-react';
 import { likeEvent, unlikeEvent } from '../services/api/event';
+import { clickNotification } from '../services/api/notification';
 
 function Stat({ iconSrc, value, altText }) {
   return (
@@ -18,6 +19,26 @@ function Stat({ iconSrc, value, altText }) {
   );
 }
 
+const getSafeImageUrl = (url) => {
+  if (!url) return '';
+  if (url.startsWith('http://')) {
+    return `https://images.weserv.nl/?url=${encodeURIComponent(url)}`;
+  }
+  return url;
+};
+
+const getTimeAgo = (notifiedAt) => {
+  const now = new Date();
+  const notifiedTime = new Date(notifiedAt);
+  const diffInMinutes = Math.floor((now - notifiedTime) / (1000 * 60));
+  if (diffInMinutes < 1) return '방금 전';
+  if (diffInMinutes < 60) return `${diffInMinutes}분 전`;
+  const diffInHours = Math.floor(diffInMinutes / 60);
+  if (diffInHours < 24) return `${diffInHours}시간 전`;
+  const diffInDays = Math.floor(diffInHours / 24);
+  return `${diffInDays}일 전`;
+};
+
 const EventCard = ({
   id,
   title,
@@ -27,17 +48,20 @@ const EventCard = ({
   likesCount,
   viewCount,
   star,
+  // notification-specific props
+  imageUrl,
+  read,
+  topicName,
+  keywordName,
+  notifiedAt,
+  clickUrl,
 }) => {
   const [cardStar, setCardStar] = useState(star);
   const [likes, setLikes] = useState(likesCount);
   const navigate = useNavigate();
 
-  const getSafeImageUrl = (url) => {
-    if (url.startsWith('http://')) {
-      return `https://images.weserv.nl/?url=${encodeURIComponent(url)}`;
-    }
-    return url;
-  };
+  const isNotification = notifiedAt !== undefined;
+  const effectiveImageUrl = imgUrl || imageUrl;
 
   const handleStarClick = async (e) => {
     e.stopPropagation();
@@ -58,10 +82,72 @@ const EventCard = ({
   };
 
   const handleCardClick = () => {
-    navigate(`/event/${id}`);
+    if (isNotification) {
+      const baseUrl = 'https://www.ajouevent.com/';
+      const trimmedUrl = clickUrl.startsWith(baseUrl)
+        ? clickUrl.replace(baseUrl, '')
+        : clickUrl;
+      clickNotification(id).catch((err) =>
+        console.error('Error clicking notification:', err)
+      );
+      navigate(`/${trimmedUrl}`);
+    } else {
+      navigate(`/event/${id}`);
+    }
   };
 
   const formattedContent = content ? content.replace(/\\n/g, ' ') : '';
+
+  if (isNotification) {
+    return (
+      <div
+        onClick={handleCardClick}
+        className={`relative flex items-start gap-3.5 w-full cursor-pointer px-5 py-4 border-b border-[#F5F6F8] hover:bg-[#FAFBFC] active:bg-[#F5F6F8] transition-colors ${
+          read ? 'bg-white' : 'bg-[#FAFCFF]'
+        }`}
+      >
+        <div className="shrink-0">
+          <div className="w-[54px] h-[54px] rounded-2xl overflow-hidden bg-[#F2F4F6]">
+            <img
+              src={getSafeImageUrl(effectiveImageUrl)}
+              alt={title}
+              loading="lazy"
+              className="w-full h-full object-cover"
+            />
+          </div>
+        </div>
+        <div className="flex flex-col flex-1 min-w-0 gap-1.5">
+          <div className="flex items-center gap-1.5 flex-wrap">
+            <div className="flex w-fit items-center gap-1 px-2 py-0.5 rounded-full bg-[#EEF3FA] text-[#003876] text-[11px] font-bold">
+              <img
+                src={`${process.env.PUBLIC_URL}/icons/notification.svg`}
+                alt="notification"
+                className="w-2.5 h-2.5 opacity-60"
+              />
+              <span>{topicName}</span>
+              {keywordName && (
+                <span className="text-[#3182F6]">· {keywordName}</span>
+              )}
+            </div>
+          </div>
+          <div
+            className={`text-[14px] font-semibold leading-snug overflow-hidden ${read ? 'text-[#333D4B]' : 'text-[#191F28]'}`}
+            style={{
+              display: '-webkit-box',
+              WebkitLineClamp: 2,
+              WebkitBoxOrient: 'vertical',
+              wordBreak: 'break-word',
+            }}
+          >
+            {title}
+          </div>
+          <div className="text-[11px] text-[#B0B8C1] font-medium">
+            {getTimeAgo(notifiedAt)}
+          </div>
+        </div>
+      </div>
+    );
+  }
 
   return (
     <div
@@ -71,7 +157,7 @@ const EventCard = ({
       <div className="w-[62px] h-[62px] flex-shrink-0 rounded-xl overflow-hidden bg-[#F2F4F6]">
         <img
           className="w-full h-full object-cover"
-          src={getSafeImageUrl(imgUrl)}
+          src={getSafeImageUrl(effectiveImageUrl)}
           alt={title}
           loading="lazy"
         />
@@ -110,5 +196,19 @@ const EventCard = ({
     </div>
   );
 };
+
+export function EventCardSkeleton() {
+  return (
+    <div className="flex items-center gap-3.5 w-full py-4 border-b border-[#F5F6F8] px-5">
+      <div className="w-[62px] h-[62px] flex-shrink-0 rounded-xl bg-[#F2F4F6] animate-pulse" />
+      <div className="flex flex-col flex-1 min-w-0 gap-1.5">
+        <div className="h-3.5 bg-[#F2F4F6] rounded-lg animate-pulse w-4/5" />
+        <div className="h-3.5 bg-[#F2F4F6] rounded-lg animate-pulse w-3/5" />
+        <div className="h-3 bg-[#F2F4F6] rounded-md animate-pulse w-2/5 mt-0.5" />
+      </div>
+      <div className="flex-shrink-0 w-8 h-8 rounded-xl bg-[#F2F4F6] animate-pulse" />
+    </div>
+  );
+}
 
 export default EventCard;
