+++ src/pages/notificationPage/NotificationPage.jsx
@@ -1,28 +1,12 @@
-import React, { useState, useEffect } from 'react';
-import { useNavigate } from 'react-router-dom';
+import React, { useState } from 'react';
 import TabBar from '../../components/layout/TabBar';
 import NotificationList from './NotificationList';
-import { getUserKeywords } from '../../services/api/subscription';
 import { readAllNotifications } from '../../services/api/notification';
 import dialog from '../../utils/dialog';
 
 const NotificationPage = () => {
-  const navigate = useNavigate();
   const [currentTab, setCurrentTab] = useState('topic');
-  const [keywordCount, setKeywordCount] = useState(0);
   const [notifications, setNotifications] = useState(0);
-
-  useEffect(() => {
-    const fetchUserKeywords = async () => {
-      try {
-        const response = await getUserKeywords();
-        setKeywordCount(response.data.length);
-      } catch (error) {
-        console.error('Error fetching user keywords:', error);
-      }
-    };
-    fetchUserKeywords();
-  }, []);
 
   const handleReadAllNotifications = async () => {
     const confirmed = await dialog.confirm('알림 읽음 처리', '정말 모든 알림을 읽음 처리할까요?');
@@ -38,8 +22,6 @@ const NotificationPage = () => {
   };
 
   const handleTabChange = (tab) => setCurrentTab(tab);
-
-  const KeywordSettingButtonClick = () => navigate('/subscribe/keywordSubscribe');
 
   return (
     <div className="flex flex-col items-center justify-start w-full min-h-screen bg-surface">
@@ -72,23 +54,6 @@ const NotificationPage = () => {
           ))}
         </div>
       </div>
-
-      {currentTab === 'keyword' && (
-        <div className="flex w-full px-5 py-3.5 bg-[#F8FAFD] justify-between items-center border-b border-[#E8EDF5]">
-          <div className="flex items-center gap-1.5">
-            <div className="w-1.5 h-1.5 rounded-full bg-[#3182F6]" />
-            <span className="text-[#333D4B] text-sm font-semibold">
-              등록된 키워드 <span className="text-[#3182F6] font-bold">{keywordCount}개</span>
-            </span>
-          </div>
-          <button
-            onClick={KeywordSettingButtonClick}
-            className="bg-[#003876] hover:bg-[#002557] active:bg-[#001a3d] text-white px-3.5 py-1.5 text-xs font-bold rounded-xl cursor-pointer border-none transition-colors"
-          >
-            키워드 설정
-          </button>
-        </div>
-      )}
 
       {currentTab === 'topic' ? (
         <NotificationList
